requirements.txt
streamlit>=1.26.0
openai<1.0.0
PyYAML
pdfminer.six
transformers
accelerate
torch
